Health goal Manager for system and user

+++ SEP490_FTCDHMM_API.Application/Dtos/RecipeDtos/CreateRecipeRequest.cs
@@ -8,7 +8,7 @@ namespace SEP490_FTCDHMM_API.Application.Dtos.RecipeDtos
         public required string Name { get; set; }
         public string? Description { get; set; } = string.Empty;
         public required string Difficulty { get; set; }
-        public double CookTime { get; set; }
+        public int CookTime { get; set; }
         public IFormFile? Image { get; set; }
         public required int Ration { get; set; }
         public required List<Guid> LabelIds { get; set; }
