+++ SEP490_FTCDHMM_API.Application/Dtos/RecipeDtos/MyRecipeResponse.cs
@@ -10,7 +10,7 @@ namespace SEP490_FTCDHMM_API.Application.Dtos.RecipeDtos
         public required string Name { get; set; } = string.Empty;
         public string? Description { get; set; } = string.Empty;
         public required DifficultyValue Difficulty { get; set; }
-        public double CookTime { get; set; }
+        public decimal CookTime { get; set; }
         public required int Ration { get; set; }
         public string? ImageUrl { get; set; }
         public required List<LabelResponse> Labels { get; set; }
